requirements.txt
python-telegram-bot==20.7
paho-mqtt==1.6.1
asyncio
